+++ tests/platforms/test_bilibili_bangumi.py
@@ -2,8 +2,8 @@ import typing
 
 import respx
 import pytest
+from httpx import Response
 from nonebug.app import App
-from httpx import Response, AsyncClient
 
 from .utils import get_json
 
@@ -11,12 +11,12 @@ if typing.TYPE_CHECKING:
     from nonebot_bison.platform.bilibili import BilibiliBangumi
 
 
-@pytest.fixture()
+@pytest.fixture
 def bili_bangumi(app: App):
-    from nonebot_bison.utils import ProcessContext
     from nonebot_bison.platform import platform_manager
+    from nonebot_bison.utils import ProcessContext, DefaultClientManager
 
-    return platform_manager["bilibili-bangumi"](ProcessContext(), AsyncClient())
+    return platform_manager["bilibili-bangumi"](ProcessContext(DefaultClientManager()))
 
 
 async def test_parse_target(bili_bangumi: "BilibiliBangumi"):