+++ tests/platforms/test_platform.py
@@ -3,7 +3,6 @@ from typing import Any
 
 import pytest
 from nonebug.app import App
-from httpx import AsyncClient
 
 now = time()
 passed = now - 3 * 60 * 60
@@ -17,7 +16,7 @@ raw_post_list_2 = raw_post_list_1 + [
 ]
 
 
-@pytest.fixture()
+@pytest.fixture
 def dummy_user(app: App):
     from nonebot_plugin_saa import TargetQQGroup
 
@@ -25,7 +24,7 @@ def dummy_user(app: App):
     return user
 
 
-@pytest.fixture()
+@pytest.fixture
 def user_info_factory(app: App, dummy_user):
     from nonebot_bison.types import UserSubInfo
 
@@ -35,7 +34,7 @@ def user_info_factory(app: App, dummy_user):
     return _user_info
 
 
-@pytest.fixture()
+@pytest.fixture
 def mock_platform_without_cats_tags(app: App):
     from nonebot_bison.post import Post
     from nonebot_bison.types import Target, RawPost
@@ -82,15 +81,15 @@ def mock_platform_without_cats_tags(app: App):
     return MockPlatform
 
 
-@pytest.fixture()
+@pytest.fixture
 def mock_platform(app: App):
     from nonebot_bison.post import Post
-    from nonebot_bison.utils import SchedulerConfig
+    from nonebot_bison.utils import Site
     from nonebot_bison.platform.platform import NewMessage
     from nonebot_bison.types import Tag, Target, RawPost, Category
 
-    class MockPlatformSchedConf(SchedulerConfig):
-        name = "mock"
+    class MockSite(Site):
+        name = "mock_site"
         schedule_type = "interval"
         schedule_setting = {"seconds": 100}
 
@@ -101,7 +100,7 @@ def mock_platform(app: App):
         is_common = True
         enable_tag = True
         has_target = True
-        scheduler = MockPlatformSchedConf
+        site = MockSite
         categories = {
             Category(1): "转发",
             Category(2): "视频",
@@ -144,20 +143,20 @@ def mock_platform(app: App):
     return MockPlatform
 
 
-@pytest.fixture()
-def mock_scheduler_conf(app):
-    from nonebot_bison.utils import SchedulerConfig
+@pytest.fixture
+def mock_site(app):
+    from nonebot_bison.utils import Site
 
-    class MockPlatformSchedConf(SchedulerConfig):
-        name = "mock"
+    class MockSite(Site):
+        name = "mock_site"
         schedule_type = "interval"
         schedule_setting = {"seconds": 100}
 
-    return MockPlatformSchedConf
+    return MockSite
 
 
-@pytest.fixture()
-def mock_platform_no_target(app: App, mock_scheduler_conf):
+@pytest.fixture
+def mock_platform_no_target(app: App, mock_site):
     from nonebot_bison.post import Post
     from nonebot_bison.types import Tag, Target, RawPost, Category
     from nonebot_bison.platform.platform import NewMessage, CategoryNotSupport
@@ -167,7 +166,7 @@ def mock_platform_no_target(app: App, mock_scheduler_conf):
         name = "Mock Platform"
         enabled = True
         is_common = True
-        scheduler = mock_scheduler_conf
+        site = mock_site
         enable_tag = True
         has_target = False
         categories = {Category(1): "转发", Category(2): "视频", Category(3): "不支持"}
@@ -211,8 +210,8 @@ def mock_platform_no_target(app: App, mock_scheduler_conf):
     return MockPlatform
 
 
-@pytest.fixture()
-def mock_platform_no_target_2(app: App, mock_scheduler_conf):
+@pytest.fixture
+def mock_platform_no_target_2(app: App, mock_site):
     from nonebot_bison.post import Post
     from nonebot_bison.platform.platform import NewMessage
     from nonebot_bison.types import Tag, Target, RawPost, Category
@@ -221,7 +220,7 @@ def mock_platform_no_target_2(app: App, mock_scheduler_conf):
         platform_name = "mock_platform"
         name = "Mock Platform"
         enabled = True
-        scheduler = mock_scheduler_conf
+        site = mock_site
         is_common = True
         enable_tag = True
         has_target = False
@@ -273,7 +272,7 @@ def mock_platform_no_target_2(app: App, mock_scheduler_conf):
     return MockPlatform
 
 
-@pytest.fixture()
+@pytest.fixture
 def mock_status_change(app: App):
     from nonebot_bison.post import Post
     from nonebot_bison.platform.platform import StatusChange
@@ -324,14 +323,14 @@ def mock_status_change(app: App):
 
 @pytest.mark.asyncio
 async def test_new_message_target_without_cats_tags(mock_platform_without_cats_tags, user_info_factory):
-    from nonebot_bison.utils import ProcessContext
     from nonebot_bison.types import Target, SubUnit
+    from nonebot_bison.utils import ProcessContext, DefaultClientManager
 
-    res1 = await mock_platform_without_cats_tags(ProcessContext(), AsyncClient()).fetch_new_post(
+    res1 = await mock_platform_without_cats_tags(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(Target("dummy"), [user_info_factory([1, 2], [])])
     )
     assert len(res1) == 0
-    res2 = await mock_platform_without_cats_tags(ProcessContext(), AsyncClient()).fetch_new_post(
+    res2 = await mock_platform_without_cats_tags(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(Target("dummy"), [user_info_factory([], [])]),
     )
     assert len(res2) == 1
@@ -345,14 +344,14 @@ async def test_new_message_target_without_cats_tags(mock_platform_without_cats_t
 
 @pytest.mark.asyncio
 async def test_new_message_target(mock_platform, user_info_factory):
-    from nonebot_bison.utils import ProcessContext
     from nonebot_bison.types import Target, SubUnit
+    from nonebot_bison.utils import ProcessContext, DefaultClientManager
 
-    res1 = await mock_platform(ProcessContext(), AsyncClient()).fetch_new_post(
+    res1 = await mock_platform(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(Target("dummy"), [user_info_factory([1, 2], [])])
     )
     assert len(res1) == 0
-    res2 = await mock_platform(ProcessContext(), AsyncClient()).fetch_new_post(
+    res2 = await mock_platform(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(
             Target("dummy"),
             [
@@ -380,14 +379,14 @@ async def test_new_message_target(mock_platform, user_info_factory):
 
 @pytest.mark.asyncio
 async def test_new_message_no_target(mock_platform_no_target, user_info_factory):
-    from nonebot_bison.utils import ProcessContext
     from nonebot_bison.types import Target, SubUnit
+    from nonebot_bison.utils import ProcessContext, DefaultClientManager
 
-    res1 = await mock_platform_no_target(ProcessContext(), AsyncClient()).fetch_new_post(
+    res1 = await mock_platform_no_target(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(Target("dummy"), [user_info_factory([1, 2], [])])
     )
     assert len(res1) == 0
-    res2 = await mock_platform_no_target(ProcessContext(), AsyncClient()).fetch_new_post(
+    res2 = await mock_platform_no_target(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(
             Target("dummy"),
             [
@@ -411,7 +410,7 @@ async def test_new_message_no_target(mock_platform_no_target, user_info_factory)
     assert "p3" in id_set_1
     assert "p2" in id_set_2
     assert "p2" in id_set_3
-    res3 = await mock_platform_no_target(ProcessContext(), AsyncClient()).fetch_new_post(
+    res3 = await mock_platform_no_target(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(Target("dummy"), [user_info_factory([1, 2], [])])
     )
     assert len(res3) == 0
@@ -419,21 +418,21 @@ async def test_new_message_no_target(mock_platform_no_target, user_info_factory)
 
 @pytest.mark.asyncio
 async def test_status_change(mock_status_change, user_info_factory):
-    from nonebot_bison.utils import ProcessContext
     from nonebot_bison.types import Target, SubUnit
+    from nonebot_bison.utils import ProcessContext, DefaultClientManager
 
-    res1 = await mock_status_change(ProcessContext(), AsyncClient()).fetch_new_post(
+    res1 = await mock_status_change(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(Target("dummy"), [user_info_factory([1, 2], [])])
     )
     assert len(res1) == 0
-    res2 = await mock_status_change(ProcessContext(), AsyncClient()).fetch_new_post(
+    res2 = await mock_status_change(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(Target("dummy"), [user_info_factory([1, 2], [])])
     )
     assert len(res2) == 1
     posts = res2[0][1]
     assert len(posts) == 1
     assert posts[0].content == "on"
-    res3 = await mock_status_change(ProcessContext(), AsyncClient()).fetch_new_post(
+    res3 = await mock_status_change(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(
             Target("dummy"),
             [
@@ -446,7 +445,7 @@ async def test_status_change(mock_status_change, user_info_factory):
     assert len(res3[0][1]) == 1
     assert res3[0][1][0].content == "off"
     assert len(res3[1][1]) == 0
-    res4 = await mock_status_change(ProcessContext(), AsyncClient()).fetch_new_post(
+    res4 = await mock_status_change(ProcessContext(DefaultClientManager())).fetch_new_post(
         SubUnit(Target("dummy"), [user_info_factory([1, 2], [])])
     )
     assert len(res4) == 0
@@ -460,13 +459,13 @@ async def test_group(
     user_info_factory,
 ):
     from nonebot_bison.types import Target, SubUnit
-    from nonebot_bison.utils import ProcessContext, http_client
     from nonebot_bison.platform.platform import make_no_target_group
+    from nonebot_bison.utils import ProcessContext, DefaultClientManager
 
     dummy = Target("dummy")
 
     group_platform_class = make_no_target_group([mock_platform_no_target, mock_platform_no_target_2])
-    group_platform = group_platform_class(ProcessContext(), http_client())
+    group_platform = group_platform_class(ProcessContext(DefaultClientManager()))
     res1 = await group_platform.fetch_new_post(SubUnit(dummy, [user_info_factory([1, 4], [])]))
     assert len(res1) == 0
     res2 = await group_platform.fetch_new_post(SubUnit(dummy, [user_info_factory([1, 4], [])]))
@@ -484,6 +483,7 @@ async def test_batch_fetch_new_message(app: App):
     from nonebot_plugin_saa import TargetQQGroup
 
     from nonebot_bison.post import Post
+    from nonebot_bison.utils import DefaultClientManager
     from nonebot_bison.platform.platform import NewMessage
     from nonebot_bison.utils.context import ProcessContext
     from nonebot_bison.types import Target, RawPost, SubUnit, UserSubInfo
@@ -538,7 +538,7 @@ async def test_batch_fetch_new_message(app: App):
     user1 = UserSubInfo(TargetQQGroup(group_id=123), [1, 2, 3], [])
     user2 = UserSubInfo(TargetQQGroup(group_id=234), [1, 2, 3], [])
 
-    platform_obj = BatchNewMessage(ProcessContext(), None)  # type:ignore
+    platform_obj = BatchNewMessage(ProcessContext(DefaultClientManager()))  # type:ignore
 
     res1 = await platform_obj.batch_fetch_new_post(
         [
@@ -570,6 +570,7 @@ async def test_batch_fetch_compare_status(app: App):
     from nonebot_plugin_saa import TargetQQGroup
 
     from nonebot_bison.post import Post
+    from nonebot_bison.utils import DefaultClientManager
     from nonebot_bison.utils.context import ProcessContext
     from nonebot_bison.platform.platform import StatusChange
     from nonebot_bison.types import Target, RawPost, SubUnit, Category, UserSubInfo
@@ -612,7 +613,7 @@ async def test_batch_fetch_compare_status(app: App):
         def get_category(self, raw_post):
             return raw_post["cat"]
 
-    batch_status_change = BatchStatusChange(ProcessContext(), None)  # type: ignore
+    batch_status_change = BatchStatusChange(ProcessContext(DefaultClientManager()))
 
     user1 = UserSubInfo(TargetQQGroup(group_id=123), [1, 2, 3], [])
     user2 = UserSubInfo(TargetQQGroup(group_id=234), [1, 2, 3], [])