+++ tests/test_context.py
@@ -5,19 +5,20 @@ from nonebug.app import App
 
 @respx.mock
 async def test_http_error(app: App):
-    from nonebot_bison.utils import ProcessContext, http_client
+    from nonebot_bison.utils import ProcessContext, DefaultClientManager, http_client
 
     example_route = respx.get("https://example.com")
     example_route.mock(httpx.Response(403, json={"error": "gg"}))
 
-    ctx = ProcessContext()
+    ctx = ProcessContext(DefaultClientManager())
     async with http_client() as client:
-        ctx.register_to_client(client)
+        ctx._register_to_client(client)
         await client.get("https://example.com")
 
     assert ctx.gen_req_records() == [
         "https://example.com Headers({'host': 'example.com', 'accept': '*/*', 'accept-encoding': 'gzip, deflate',"
-        " 'connection': 'keep-alive', 'user-agent': 'Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like"
-        " Gecko) Chrome/51.0.2704.103 Safari/537.36'}) | [403] Headers({'content-length': '15', 'content-type':"
+        " 'connection': 'keep-alive', 'user-agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36"
+        " (KHTML, like Gecko) Chrome/130.0.0.0 Safari/537.36 Edg/130.0.0.0'}) | [403] Headers({'content-length': '"
+        "15', 'content-type':"
         ' \'application/json\'}) {"error": "gg"}'
     ]