+++ tests/utils.py
@@ -89,6 +89,7 @@ add_reply_on_id_input_search = (
 
 
 class BotReply:
+
     @staticmethod
     def add_reply_on_platform(platform_manager, common_platform):
         return (
@@ -146,6 +147,10 @@ class BotReply:
         extra_text = ("1." + target_promot + "\n2.") if target_promot else ""
         return extra_text + base_text
 
+    @staticmethod
+    def add_reply_platform_unavailable(platform: str, reason: str) -> str:
+        return f"无法订阅 {platform}，{reason}"
+
     add_reply_on_id_input_error = "id输入错误"
     add_reply_on_target_parse_input_error = "不能从你的输入中提取出id，请检查你输入的内容是否符合预期"
     add_reply_on_platform_input_error = "平台输入错误"
@@ -154,3 +159,34 @@ class BotReply:
     )
     add_reply_on_tags_need_more_info = "订阅标签直接输入标签内容\n屏蔽标签请在标签名称前添加~号\n详见https://nonebot-bison.netlify.app/usage/#%E5%B9%B3%E5%8F%B0%E8%AE%A2%E9%98%85%E6%A0%87%E7%AD%BE-tag"
     add_reply_abort = "已中止订阅"
+    no_permission = "您没有权限进行此操作，请联系 Bot 管理员"
+
+    @staticmethod
+    def add_reply_on_add_cookie(platform_manager, common_platform):
+        from nonebot_bison.utils.site import is_cookie_client_manager
+
+        return (
+            "请输入想要添加 Cookie 的平台，目前支持，请输入冒号左边的名称：\n"
+            + "".join(
+                [
+                    f"{platform_name}: {platform_manager[platform_name].name}\n"
+                    for platform_name in common_platform
+                    if is_cookie_client_manager(platform_manager[platform_name].site.client_mgr)
+                ]
+            )
+            + "要查看全部平台请输入：“全部”\n中止添加cookie过程请输入：“取消”"
+        )
+
+    @staticmethod
+    def add_reply_on_add_cookie_input_allplatform(platform_manager):
+        from nonebot_bison.utils.site import is_cookie_client_manager
+
+        return "全部平台\n" + "\n".join(
+            [
+                f"{platform_name}: {platform.name}"
+                for platform_name, platform in platform_manager.items()
+                if is_cookie_client_manager(platform.site.client_mgr)
+            ]
+        )
+
+    add_reply_on_input_cookie = "请输入 Cookie"